BorderControl
Can receive an unknown amount of lines from the console until the command "End" is received, on each line, there will be a piece of information for either a citizen or a robot who tries to enter your city. After the "End" command on the next line, you have to write a single number representing the last digits of fake ids. The output of program consist all of detained Ids each on a separate line in the order of input.