BirthdayCelebrations
Can receive from the console an unknown number of lines. Until the command "End" is received, each line will contain information in one of the following  Citizen, Robot, Pet. After the "End" command on the next line, you have to write a single number representing a specific year, and will print all birthdates (of both Citizen and Pet) in that year in the format day/month/year in the order of input.